# server/pkg/publicapi/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Multica Public API
  version: v1
  description: >-
    Canonical v1 resource contract shared by Multica API surfaces. The Plugin
    API is the first deployed consumer and applies installation identity plus
    manifest scopes. Plugin-specific context and storage remain extensions.
servers:
  - url: https://plugin-api.multica.ai/v1
    description: Installed Plugin and invocation callback tokens
security:
  - pluginBearer: []
paths:
...
  /issues/{issue_ref}:
    parameters:
      - $ref: "#/components/parameters/IssueRef"
    get:
      operationId: getIssue
      summary: Get an issue
      x-multica-contract: shared_resource
      x-multica-scope: issues:read
      responses:
        "200":
          description: Issue
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      operationId: patchIssue
      summary: Update an issue title or description
      x-multica-contract: shared_resource
      x-multica-scope: issues:write
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchIssueRequest"
      responses:
        "200":
          description: Updated issue
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    pluginBearer:
      type: http
      scheme: bearer
      bearerFormat: mpi_ or mpc_ token
  parameters:
    IssueRef:
      name: issue_ref
      in: path
      required: true
      description: Issue UUID or workspace identifier such as MUL-123
      schema:
        type: string
...
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: Weak or strong ETag containing the expected positive revision
      schema:
        type: string
...
  headers:
    ETag:
      description: Weak ETag derived from the resource revision
      schema:
        type: string
  responses:
    Problem:
      description: Stable API problem
      headers:
        X-Request-Id:
          description: Request correlation identifier
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, detail, request_id, error]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        request_id: { type: string }
        error:
          type: string
          deprecated: true
          description: Compatibility alias for detail
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
...
    Issue:
      type: object
      required:
        - id
        - workspace_id
        - number
        - identifier
        - title
        - description
        - status
        - priority
        - assignee_type
        - assignee_id
        - creator_type
        - creator_id
        - parent_issue_id
        - project_id
        - position
        - stage
        - start_date
        - due_date
        - created_at
        - updated_at
        - revision
        - last_activity_at
        - metadata
        - properties
      properties:
        id: { type: string }
        workspace_id: { type: string }
        number: { type: integer, format: int32 }
        identifier: { type: string }
        title: { type: string }
        description: { type: [string, "null"] }
        status: { type: string }
        status_category: { type: string }
        priority: { type: string }
        assignee_type: { type: [string, "null"] }
        assignee_id: { type: [string, "null"] }
        creator_type: { type: string }
        creator_id: { type: string }
        parent_issue_id: { type: [string, "null"] }
        project_id: { type: [string, "null"] }
        position: { type: number }
        stage: { type: [integer, "null"], format: int32 }
        start_date: { type: [string, "null"], format: date }
        due_date: { type: [string, "null"], format: date }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        revision: { type: integer, format: int64, minimum: 1 }
        last_activity_at: { type: [string, "null"], format: date-time }
        metadata: { type: object, additionalProperties: true }
        properties: { type: object, additionalProperties: true }
    PatchIssueRequest:
      type: object
      properties:
        expected_revision: { type: integer, format: int64, minimum: 1 }
        title: { type: string }
        description: { type: string }
      anyOf:
        - required: [title]
        - required: [description]
...